Components: Move PostStatsCard to client
