+++ client/components/post-stats-card/stories/index.stories.jsx
@@ -1,6 +1,6 @@
 import PostStatsCard from '../';
 
-export default { title: 'packages/components/Post Stats Card' };
+export default { title: 'client/components/Post Stats Card' };
 
 const PostStatsCardVariations = ( props ) => (
 	<PostStatsCard
